nrt: split logging helpers
Split logging helpers in their own file.
Trivial code movement, no code changes.
