+++ pkg/noderesourcetopology/pluginhelpers.go
@@ -18,7 +18,6 @@ package noderesourcetopology
 
 import (
 	"context"
-	"encoding/json"
 	"fmt"
 	"time"
 
@@ -127,25 +126,4 @@ func newScoringHandlers(strategy scoreStrategy, resourceToWeightMap resourceToWe
 			return containerScopeScore(pod, zones, strategy, resourceToWeightMap)
 		},
 	}
-}
-
-func logNumaNodes(desc, nodeName string, nodes NUMANodeList) {
-	for _, numaNode := range nodes {
-		numaLogKey := fmt.Sprintf("%s/node-%d", nodeName, numaNode.NUMAID)
-		klog.V(6).InfoS(desc, stringify.ResourceListToLoggable(numaLogKey, numaNode.Resources)...)
-	}
-}
-
-func logNRT(desc string, nrtObj *topologyv1alpha1.NodeResourceTopology) {
-	if !klog.V(6).Enabled() {
-		// avoid the expensive marshal operation
-		return
-	}
-
-	ntrJson, err := json.MarshalIndent(nrtObj, "", " ")
-	if err != nil {
-		klog.V(6).ErrorS(err, "failed to marshal noderesourcetopology object")
-		return
-	}
-	klog.V(6).Info(desc, "noderesourcetopology", string(ntrJson))
 }
